Split apps for testability
