+++ server/app.ts
@@ -1,0 +1,3 @@
+import * as express from "express"
+
+export const app = express()
