+++ server/index.ts
@@ -1,3 +1,5 @@
+import * as express from "express"
+import * as path from "path"
 import { Message, User, TwitterIntegration, Connection } from "./models"
 
 require('dotenv').config();
@@ -13,7 +15,10 @@ for (const model of [Message, TwitterIntegration, User]) {
 // Connection must be cleaned up after restart server
 Connection.sync({force: true})
 
-import { app } from "./server";
+import { app } from "./app"
+app.use(express.static(path.join(__dirname, '../../ui/dist')))
+import "./auth"
+import "./api"
 import "./wsserver"
 
 app.listen(8080);
